Finalize first implementation of layer uploads to geonode
- delete temp dir
- move uploader task closer to its callers
