+++ pyproject.toml
@@ -21,6 +21,8 @@ pytest-xvfb = "^2.0.0"
 Pygments = "^2.8.0"
 pymdown-extensions = "^8.1.1"
 pytest-custom-exit-code = "^0.3.0"
+mypy = "^0.930"
+qgis-stubs = "^0.1.0"
 
 [tool.poetry.dev-dependencies]
